Correction de warning

+++ Api_c_sharp/Api_c_sharpTests/ControllersTestsUnitaire/AnnonceControllerTests.cs
@@ -1,26 +1,16 @@
 ﻿using Api_c_sharp.Controllers;
-using Api_c_sharp.Controllers;
 using Api_c_sharp.Mapper;
 using Api_c_sharp.Models.Entity;
 using Api_c_sharp.Models.Repository;
 using Api_c_sharp.Models.Repository.Interfaces;
-using Api_c_sharp.Models.Repository.Managers;
 using Api_c_sharp.Models.Repository.Managers.Models_Manager;
 using AutoMapper;
 using AutoPulse.Shared.DTO;
-using Google.Apis.Util;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging.Abstractions;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Numerics;
 using System.Security.Claims;
-using System.Threading.Tasks;
 
 namespace Api_c_sharp.ControllersUnitaires.Tests
 {
@@ -671,7 +661,7 @@ namespace Api_c_sharp.ControllersUnitaires.Tests
         public async Task GetByMiseEnavant()
         {
             // Act
-            var result = await _controller.GetByIdMiseEnAvant((int)_objetcommun.IdMiseEnAvant);
+            var result = await _controller.GetByIdMiseEnAvant(_objetcommun.IdMiseEnAvant);
             // Assert
             Assert.IsNotNull(result);
             Assert.IsNotNull(result.Value);
